Initial commit of OpenJS Foundation artwork repo
This is the initial commit of the OpenJS Foundation artwork repo.  Thank you to
everyone who helped me track down vector versions of project logos.  The goal is
to keep this repository in sync with the various artwork assets used by
projects.

Contributions to this repository are welcomed, particularly if you see a gap.
Guidelines for contributing are in the README.

Please be aware that these marks are covered by the OpenJS Foundation trademark
policy, as adopted by the Board of Directors.  Please read and understand the
policy before using a mark: https://trademark-policy.openjsf.org

If you need help converting or modifying a vector version of your artwork,
please reach out to operations@openjsf.org.

Signed-off-by: Brian Warner <brian@bdwarner.com>