Add first round of hex stickers
This commit includes artwork for the first round of hex stickers.  These
versions are mostly the stacked (vertical) version of each project's logo.  We
had recently gotten feedback that stickers without project names could be a
little obscure, so we're trying to see if these help (except where we've been
explicitly told otherwise).

Projects are welcome to edit their stickers.  Each time we need to create
another batch, we'll pull the latest art from this repo.

Signed-off-by: Brian Warner <brian@bdwarner.com>