Add member logos
This commit establishes a directory of member logos, which can be linked from
the website and from Landscapeapp.

Trademarks and logos not indicated on the list of OpenJS Foundation trademarks
are trademarks or registered trademarks of their respective holders. Use of
them does not imply any affiliation with or endorsement by them.

The list of OpenJS Foundation trademarks may be found here:
https://trademark-list.openjsf.org

Signed-off-by: Brian Warner <brian@bdwarner.com>